feat: update fink to new vike v1
