+++ inlang/source-code/editor/package.json
@@ -93,7 +93,7 @@
 		"tailwindcss": "^3.3.0",
 		"typescript": "5.2.2",
 		"unplugin-icons": "^0.15.0",
-		"vike": "0.4.144",
+		"vike": "0.4.152",
 		"vite": "4.5.0",
 		"vite-plugin-node-polyfills": "0.12.0",
 		"vite-plugin-solid": "2.7.0",
