+++ inlang/source-code/editor/src/renderer/+config.h.ts
@@ -1,0 +1,5 @@
+import type { Config } from "vike/types"
+
+export default {
+	clientRouting: true,
+} satisfies Config
