+++ inlang/source-code/editor/src/services/auth/src/pages/auth-callback/+route.ts
